requirements.txt
google-cloud-bigquery==3.0
dbt-bigquery==1.4.0
pandas==1.3.0
numpy==1.21.6
db-dtypes==1.1.0